Title

Title

Usage

facet_barchart(.data, xcol, ycol, groupcol, textsize = 1)

Arguments

.data

A dataframe containing your data, ideally a 3 column dataframe with one value column and two categorical columns

xcol

The column which contains the variable to go on the x axis of the column chart

ycol

The column which contains the value to go on the y axis

groupcol

The column which contains the categorical variable which the graphs will be faceted by

textsize

The text size, initially set to 1, the size is a ratio to the normal dluhc_style size of 2

Value

a ggplot object

Examples

df <- dluhctheme::Net_Additions_Regional
df <- dplyr::filter(df,Regions %in% c("North East","North West","South East","South West","London", "East Midlands"))
#> Error in dplyr::filter(df, Regions %in% c("North East", "North West",     "South East", "South West", "London", "East Midlands")): ℹ In argument: `Regions %in% ...`.
#> Caused by error in `Regions %in% c("North East", "North West", "South East", "South West", "London",
#>     "East Midlands")`:
#> ! object 'Regions' not found
df$year <- substr(df$Year,7,10)
facet_barchart(df,xcol = year,ycol = Net_Additions,groupcol=Region)
#> Error in palette(...): Insufficient values in manual scale. 9 needed but only 6 provided.